samples: philosophers: remove redundant items in prj.conf
The normal and posix philosophers samples seemed to have
a number of redundant choices in prj.conf.

Let's reduce to the minimum required for the sample. The
CONFIG_DEBUG_THREAD_INFO option is part of the
documentation, so leave it as-is.

(cherry picked from commit dd33a8152a36db9e624859a08ae5eb1676274b78)

Original-Signed-off-by: Christopher Friedt <cfriedt@meta.com>
GitOrigin-RevId: dd33a8152a36db9e624859a08ae5eb1676274b78
Change-Id: Ibfbedcebbd8c20fbec8b1dcc352215d6ab024ff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5280064
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>