Bluetooth: BAP: Update log for unicast iso recv without endpoint
Since we always set up the ISO data path for endpoints in both
directions (due to limitations in the ISO API), we can actually
receive valid (empty) SDUs on a CIS in a direction that has not
been configured.

This is not the ideal solution, but prevents unncessary LOG_ERR.
The ideal solution is to modify the ISO API to provide the
flexibility that BAP needs w.r.t. ISO data paths, but that is
a larger change for later.

(cherry picked from commit 13a357b0198defde8f5c34920bec25b9ea2a0855)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 13a357b0198defde8f5c34920bec25b9ea2a0855
Change-Id: Ic765c95970bb646b4eb7c407da7c3a027ec5a299
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5280029
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>