libc: Move gmtime_r into common
gmtime_r() has been in the minimal libc for years, however it was not
added to expcetions due to an overlook. In order to do this however, it
has to be moved first to the common libc area, so that it's available
to any libc that may not implement it.

(cherry picked from commit 0966be01fc578c790ea6cd1f96b39b094fc1284e)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: 0966be01fc578c790ea6cd1f96b39b094fc1284e
Change-Id: I34659141d203f6bb6a0d2e8cc7fa2c7bf5af9006
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5333035
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>