Bluetooth: Host: Set `conn->err` in prio
The goal is to not access `buf` in the non-prio `hci_disconn_complete`
so that the buffer does not need to be retained.

This commit moves the copy of `conn->err` to the prio handler.

(cherry picked from commit dd7caf667c8edd08e841bfb14f65ff96275d232c)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: dd7caf667c8edd08e841bfb14f65ff96275d232c
Change-Id: Iac3e59b1cba5d0b4b9faa2f8c5ecb9a1141bdd05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5327024
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>