drivers: ethernet: adin2111: always append crc32 to the eth frame
Enable appending of a crc32 at the end of the frame by the MAC,
always. This is always needed since the driver is not adding it.

This field has nothing to do with Generic SPI protocol-related
8-bit CRC, so this patch removes the CONFIG_ETH_ADIN2111_SPI_CFG0
choice related to this setting.

Testing without this flag set, packets are not forwareded in the
network, since the driver is not appending any crc32 header
to the frame.

(cherry picked from commit 5784a3a78ff654f88076e32f52dec28b631b00dc)

Original-Signed-off-by: Angelo Dureghello <adureghello@baylibre.com>
GitOrigin-RevId: 5784a3a78ff654f88076e32f52dec28b631b00dc
Change-Id: If09d2932e6b7c402b5c29c489ef62927b72522c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5359114
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>