drivers: dma: dma_dw_common: Disable channel even if draining times out
If the channel suspend with draining fails on stop because of reasons
outside of the scope of the DMA driver (the peripheral is powered off
before trying to drain for example) we must continue and disable the
channel.

The channel can be released by the client despite of it remained enabled.
A new DMA channel request can pick the channel (as it is released) but
re-configuration is going to be skipped and the use of the channel is going
to fail. Then we will see the same drain timeout on channel stop again
since the channel retained the configuration which resulted the first
timeout.

The drain timeout was made fatal by an earlier commit which fixed the
WAIT_FOR return value handling.

(cherry picked from commit 84631ce17c9c3a1935befe0b9e669696f0184105)

Original-Fixes: 6226f9e6e44f ("dma: dw: fix the return value check")
Original-Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
GitOrigin-RevId: 84631ce17c9c3a1935befe0b9e669696f0184105
Change-Id: Iedc04a4f8189e307e8f618839316376e7c36cf82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5281895
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>