drivers: dma: dma_nxp_edma: query channel number from HAL config
Currently, the number of channels supported by the controlled
is computed based on the size of the channel array. This
works well only if there's no gaps (i.e: "dma-channels" property
is used or "valid-channels" property is used with contiguous
channels) but will break if there are any gaps. For instance,
if the user wants to use channels 16 and 17 and specifies them
through the "valid-channels" property, they won't be allowed
to do so because dma_request_channels() will stop at channel 1.
As such, to fix this, simply use the number of channels from
the HAL configuration which is the maximum number of channels.

(cherry picked from commit c07d9874312419eab25937eb203cdeb54e52ab6b)

Original-Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
GitOrigin-RevId: c07d9874312419eab25937eb203cdeb54e52ab6b
Change-Id: Ib6920f0b641ace67fe8db39ee06c706436b96a10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5283285
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>