drivers: ipm: esp32: Allow doorbell without data transfer
IPM drivers are commonly used to send notifications/cause interrupts
without any transfer of data. To add this use case in the ESP32 IPM
driver, the guard statement is appended so that the pointer to the
data buffer is allowed to be zero only if the size of the data to be
transferred is zero. If size is given as 0 and data is equal to NULL,
we are thus only using the IPM as a doorbell, not to transfer data.

(cherry picked from commit 83192c71e413acd795fd9082b3b8ac59d199cbf6)

Original-Signed-off-by: Celina Sophie Kalus <hello@celinakalus.de>
GitOrigin-RevId: 83192c71e413acd795fd9082b3b8ac59d199cbf6
Change-Id: I15620454d017e5c614324a55dc3776ab263c1c58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5327037
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>