hostap: Reserve the heap
By default (-1) libc uses all the available heap, this is good to
efficiently use the RAM, but for scenarios which there isn't enough RAM
the build will be success, but we see runtime failures.

Reserved the necessary RAM based on tests ahead to catch such scenarios,
of course this would mean that we might be no using full RAM in case
more is left, but that's never the case the with Wi-Fi :).

And this also adds the constant evaluation and fine tuneing like we do
for kernel heap and other parameters.

(cherry picked from commit 2f29c1057ffe7cd48406038129cf70743eaae850)

Original-Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>
Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 2f29c1057ffe7cd48406038129cf70743eaae850
Change-Id: Ia1caf4c17f170f802c0142e4693612791c08e036
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5428728
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>