drivers: serial: uart_async_rx: Optimize RAM usage
Since there is only one consumer of the data stored in the buffers,
it is enough to have one read index variable which can be stored in
the data associated with the module and not in the buffer space (where
there is a read index for each buffer).

Additionally, we can safely assume that module works with small buffers
so 127 byte limit is enough. Based on that assumption completed flag
can be stored on a single byte together with write index. After this
change, control data for each buffer takes 1 byte (3 bytes previously).

(cherry picked from commit 8bc5111c27df233520b6a94f5041f08fcd887fcb)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 8bc5111c27df233520b6a94f5041f08fcd887fcb
Change-Id: I859c645f28a783207624f92a621b2d5382e92025
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5400965
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>