xtensa: exiting simulator on fatal errors
There was old code to exit the simulator when there is a fatal
error. This updates the #ifdef and the function name so that
it can be used to exit both QEMU and simulator when a fatal
error occurs. This should help with twister runs as this will
fail immediately instead of having wait for timeout.

(cherry picked from commit d89e7f63fd4d28fbfe54c418b7bea41602d76212)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: d89e7f63fd4d28fbfe54c418b7bea41602d76212
Change-Id: I88caa6aadba32b78609b637494fffd4d347b8de9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5389933
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>