pm: Fix a corner case when policy returns NULL
When the policy returns NULL pm_system_suspend was assuming that
the current state in z_cpus_pm_state was ACTIVE, since that is the
state set after the core wakes and return to this function. The
problem is that in cases where the cpu does not preserve the context,
and returns to this function, z_cpus_pm_state has the value of the
last state used and the cpu use it again.

Fix it setting z_cpus_pm_state to ACTIVE every time the policy returns
NULL.

(cherry picked from commit e2058a953707c93cab249575db67aecd184b3465)

Original-Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: e2058a953707c93cab249575db67aecd184b3465
Change-Id: I4e898da56e47ecc58f34e3e3708d43949027a604
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5385352
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>