drivers: ieee802154: gracefully handle invalid Ack timestamp
The nRF IEEE 802.15.4 driver might report a received Ack frame with
invalid timestamp, if the timestamp could not have been taken. The upper
layers are not prepared to handle such a case as they expect that for a
received frame, the timestamp is always present and valid.

This commit detects this situation and handles it gracefully by
reporting the transmission as failed as if no Ack was received.

(cherry picked from commit 774dd165557e71bb2d3e67dd6c62d8229c1017be)

Original-Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>
GitOrigin-RevId: 774dd165557e71bb2d3e67dd6c62d8229c1017be
Change-Id: I74efd61b27621d4504db905db2e25d60dca7a590
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5439938
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>