ci: twister: Add `--specs` to ccache ignore option list
This commit adds the compiler `--specs=*` flag to the ccache ignore option
list because ccache is unable to resolve the toolchain-provided specs file
path and will consider source files to be uncacheable if it is unable to
read the specified specs file.

Note that adding `--specs=*` to the ignore option list is not a problem
because it is unlikely for the content of the toolchain libc spec file to
change without the compiler executable itself changing.

(cherry picked from commit d3f9f391ad468e18311ce351a259e59248125d9b)

Original-Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: d3f9f391ad468e18311ce351a259e59248125d9b
Change-Id: Ifaaca3879830978d8939e5e3c249f2f6f411fffa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5351972
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>