dts: nordic: Add DPPIC and IPCT nodes to nRF54H20 EngA
And add the corresponding bindings.

Move alse the already existing bindings for nrf-dppic, nrf-ppi,
and nrf-ipc so they are located together with the new ones and
in more appropriate folders (DPPIC and PPI peripherals are not
related to ARM, and IPC is for sending and receiving events,
not messages, so ipm/ does not seem to be the best fit for it).

(cherry picked from commit 796d09d2a650c3726ea9569397119dae5d2c7711)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 796d09d2a650c3726ea9569397119dae5d2c7711
Change-Id: Ie75b3b55386e6afa49945d2d12f9e5bb0f653b78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5368136
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>