Bluetooth: Audio: Scan Delegator: Guard for duplication in add_source
Added a check for duplication of add source operation in the Scan
Delagator. BAP 6.5.4 states that an assistant may not perform add
source operation if it results in a duplication of a source. While
this is not mentioned for the Scan Delegator, the spirit of the spec.
must be that no duplications can exists in the recv_state, which is
what this commit fixes.

(cherry picked from commit 0fe8ddd75e151d3d779011df4e3914c97add74c0)

Original-Signed-off-by: Fredrik Danebjer <frdn@demant.com>
GitOrigin-RevId: 0fe8ddd75e151d3d779011df4e3914c97add74c0
Change-Id: I52047c466092e0062b8530feea690b9156a9ca51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5402699
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>