net: dns: Split long resolving to smaller pieces
If getaddrinfo() is called with a long DNS timeout, then split
the timeout to smaller pieces with exponential backoff. Reason
for this is that if a DNS query is lost, then we do not need
to wait for a long time to find it out.

(cherry picked from commit 5be84131d6f8f68126d58134a7aa307b93050acb)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 5be84131d6f8f68126d58134a7aa307b93050acb
Change-Id: Ic284eaca15a0b3fb21d707850b286176048280bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5351946
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>