net: zperf: Fix multicast data failure to receive server's AckFIN packet
In iperf2, for multicast data, if it is a client, it will not wait for
the server's AckFIN packet. Because the iperf2 server will not send an
AckFIN packet.
So in zperf_upload_fin(), an error will occur when the zperf client
waits for the server's AckFIN packet.
Multicast only send the negative sequence number packet and doesn't
wait for a server ack can fix this issue.

(cherry picked from commit 03fce7dd3cdccef4538745ac9301a10557f483f2)

Original-Signed-off-by: Gang Li <gang.li_1@nxp.com>
GitOrigin-RevId: 03fce7dd3cdccef4538745ac9301a10557f483f2
Change-Id: Id60f3797f2da790bc0abc518e2af6d111e907f75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5356026
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>