ci: codecov: Add `--specs` to ccache ignore option list
This commit adds the compiler `--specs=*` flag to the ccache ignore option
list because ccache is unable to resolve the toolchain-provided specs file
path and will consider source files to be uncacheable if it is unable to
read the specified specs file.

Note that adding `--specs=*` to the ignore option list is not a problem
because it is unlikely for the content of the toolchain libc spec file to
change without the compiler executable itself changing.

(cherry picked from commit ab9f6b456b6daa472a6430bd9b80f28326cb8cc9)

Original-Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: ab9f6b456b6daa472a6430bd9b80f28326cb8cc9
Change-Id: Id422b0bbae37c1b999014ced6191540c7cb5536c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5366711
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>