doc/develop api_lifecycle: Mention adding the arch review label
For breaking API changes, let's mention explicity the PR should
be labelled with the "Architecture Review" label, as that is
the one we base the automation on.
As otherwise it is not clear enough for developers that they
should add it, and without it (and the automation surrounding it),
it is too easy for a PR to be merged by mistake.

(cherry picked from commit e1b7b450c82cd8236a36271fab49cbb9067a9d22)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: e1b7b450c82cd8236a36271fab49cbb9067a9d22
Change-Id: If1c6ef014f14fa1d0c4b132166f1b2da5bc68c98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5370931
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>