llext: arch_elf_relocate: pass opval unmodified
The opval argument of arch_elf_relocate() was modified by
adding the value stored at opaddr before passing it to
arch_elf_relocate(). This presumed that the addend would
always be stored as a raw value at opaddr, which is not the
case for all relocation types.

This PR modifies opval to be the absolute address of opval,
and moves the addition of the addend from llext_link_plt()
to the implementation of arch_elf_relocate().

(cherry picked from commit 5cac834bb65113bb93e9a5546bf75732c36f1ddb)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki@arge-andreasen.me>
GitOrigin-RevId: 5cac834bb65113bb93e9a5546bf75732c36f1ddb
Change-Id: Ie970e43acd8b69357502f739b9af30436826e459
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5374398
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>