cmake: extensions: Change zephyr_file() argument parser prefix
Replace `FILE` with `ZFILE` as the prefix for `cmake_parse_arguments()`.
This is because `zephyr_file()` accepts `SUFFIX` as function argument,
and it should not be accessible as `FILE_SUFFIX` as it would conflict
with the CMake cache variable of the same name. Accessing the argument
as `ZFILE_SUFFIX` prevents the cached value from slipping into the
function scope.

(cherry picked from commit 50f0454fbd446b5bb0653b25cedccd022aa18006)

Original-Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>
GitOrigin-RevId: 50f0454fbd446b5bb0653b25cedccd022aa18006
Change-Id: Ifa5ad3af5862c8bdb36bb5ad71214344f0b26c9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5388020
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>