Bluetooth: CAP: Fix issue with codec_cfg in CAP uni start
The codecs used for the bap bt_bap_stream_config was only
valid for the lifetime of the procedure, which meant
that by the end of the procedure the stream->codec_cfg
became invalid.

This is fixed by using the pointer provided to the
CAP API, and documentating the lifetime of the codec_cfg.

(cherry picked from commit 62ac753a6148b7b5e381bd86f72174b82b2d99c5)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 62ac753a6148b7b5e381bd86f72174b82b2d99c5
Change-Id: I9a9eb8bbe833b72c135893ebf438d15fde856d79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5380377
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>