samples/basic/hash_map: Remove clarification sentence
This current sentence would need to be clarified a bit
further to be really correct:
It is only when building with the host C library that it
applies, but with native_sim we can build with other
libC's.
The twister tests definitions are testing only with
embedded C libraries,
and there is anyway no harm if somebody defines the heap
size even if using the host libC.
So as such it would seem better to just remove this
sentence.

(cherry picked from commit e807b7e1c2f261a30e349a8bd297f36ef6f862bb)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: e807b7e1c2f261a30e349a8bd297f36ef6f862bb
Change-Id: Iee5bd9607203e0c18b8f74df7cc36a83837f9786
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5376203
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>