ci: doc-build: increase timeout for doc build jobs
With the introduction of the new runners, the time it takes to build the
documentation in CI has slowed down to the point that the job often times
out.

Applied a conservative 2x factor to the timeouts that were in place.

(cherry picked from commit 107b335cce0eac5577b74df47bbeacb46f2dd165)

Original-Signed-off-by: Benjamin Cabé <benjamin@zephyrproject.org>
GitOrigin-RevId: 107b335cce0eac5577b74df47bbeacb46f2dd165
Change-Id: I888675c439651912e0e7f8d7217ff195cee86d2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5383207
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Yuval Peress <peress@google.com>